Bug 27552: Use bundled dir on CentOS/RHEL 6
On CentOS/RHEL 6, add the bundled directory from the firefox package to
the LD_LIBRARY_PATH.

In addition, we don't reset anymore the previous LD_LIBRARY_PATH value,
but only add new directories to it. This allows users to set LD_LIBRARY_PATH
before starting the browser if additional libraries are needed.