Preparing a build2 for 8.0.1
We pick up three patches that are planned for Firefox 60.2.1esr but we
can't wait for the final release candidate as we run out of time before
Mexico.

Furthermore, back out the moat related fix in Tor Launcher. We want to
give it a bit more testing time on alpha.