Bug 29235: Build our own version of python3.6
The build of HTTPS-Everywhere requires python3.6, however the python3.6
package which we were using has now been removed from Debian Unstable.

We now build python 3.6 ourselves. As the different platforms use a
different suite by default, we add the common-stretch target to
rbm.conf, which we use to avoid building a different python for each
platform.