Bug 32053: Fix LLVM reproducibility issues
This patch contains two backports for reproducibility issues found and
fixed in LLVM's optimization code.

The patch from bug 42574 might not be necessary in our context as I only
witnessed problems when compiling Stylo with an older Rust (1.32.0) and
then omptimizing with LLVM 8. However, it's hard to exclude any
mismatching builds we encountered so far from being affected by that
issue, so I think it is safer to backport that patch as well.

Bug 43909 is an issue we and Mozilla definitely hit and very likely
responsible for at least the vast majority of different builds we got
over the course of the last weeks and months. Many thanks to Alex
Crichton for the invaluable help in tracking this issue down.